+++ cypress/e2e/finance/fiscalYears/utc-is-displayed-when-create-or-view-fiscal-year.cy.js
@@ -31,6 +31,7 @@ describe('ui-finance: Fiscal Year', () => {
     { tags: ['criticalPath', 'thunderjet'] },
     () => {
       FiscalYears.createDefaultFiscalYear(defaultFiscalYear);
+      FiscalYears.closeThirdPane();
       FiscalYears.searchByName(defaultFiscalYear.name);
       FiscalYears.selectFisacalYear(defaultFiscalYear.name);
       FiscalYears.checkCreatedFiscalYear(defaultFiscalYear.name);