+++ cypress/support/fragments/inventory/inventorySearchAndFilter.js
@@ -804,6 +804,7 @@ export default {
     cy.wait(ONE_SECOND);
     holdingsPermanentLocationAccordion.find(TextField()).click();
     cy.do(holdingsPermanentLocationAccordion.find(Checkbox(location)).click());
+    cy.wait(ONE_SECOND);
   },
 
   checkRowsCount: (expectedRowsCount) => {
@@ -883,7 +884,6 @@ export default {
 
   browseOptionsDropdownIncludesOptions(options) {
     const browseOptionsDropdown = Select('Search field index');
-    this.switchToBrowseTab();
     options.forEach((name) => {
       cy.expect(browseOptionsDropdown.has({ content: including(name) }));
     });
@@ -900,7 +900,7 @@ export default {
   verifySearchAndFilterPane() {
     this.searchTabIsDefault();
     this.instanceTabIsDefault();
-    this.searchTypeDropdownDefaultValue('all');
+    this.searchTypeDropdownDefaultValue('Keyword (title, contributor, identifier, HRID, UUID)');
     this.verifySearchFieldIsEmpty();
     cy.expect([
       searchToggleButton.exists(),
@@ -936,7 +936,7 @@ export default {
   },
 
   searchTypeDropdownDefaultValue(value) {
-    cy.expect(searchTypeDropdown.has({ value }));
+    cy.expect(searchTypeDropdown.has({ checkedOptionText: value }));
   },
 
   verifySearchFieldIsEmpty() {